infrastructure/modules/ingestion/lambdas/theoddsapi/requirements.txt
snowflake-connector-python==4.2.0
